fixed some tests
open todos:
- add Snowflake type + make it compatible with serde and sea-orm
- use system time instead of chrono to generate timestamp (~70 % faster than chrono)
- make snowflake ids serialize to a string
