+++ README.md
@@ -26,7 +26,7 @@ With its intuitive interface and powerful features, Financrr empowers users to t
 
 - [Docs](https://financrr.github.io/financrr-app/docs/)
 - [Frontend stage](https://financrr.github.io/financrr-app/frontend-stage/)
-- [Backend stage](https://api-stage.financrr.app/swagger-ui/)
+- [Backend stage](https://api-stage.financrr.app/openapi/swagger-ui/)
 
 ## 🔧 Contributing
