+++ backend/backend/Cargo.toml
@@ -44,10 +44,12 @@ redis = {  version = "0.25.4", features = ["aio", "tokio-comp"] }
 # Serialization and deserialization
 serde.workspace = true
 serde_json = "1.0.118"
+serde_yml = "0.0.10"
 
 # Generating openapi spec
 utoipa = { version = "4.2.3", features = ["actix_extras", "openapi_extensions", "time", "preserve_order", "preserve_path_order", "non_strict_integers"] }
 utoipa-swagger-ui = { version = "7.1.0", features = ["actix-web", "debug-embed", "reqwest"] }
+utoipa-scalar = { version = "0.1", features = ["actix-web"] }
 utoipauto = { version = "0.1.11", features = ["generic_full_path"] }
 
 # Logging
