+++ backend/backend/src/api/routes/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod account;
 pub(crate) mod budget;
 pub(crate) mod currency;
+pub(crate) mod openapi;
 pub(crate) mod session;
 pub(crate) mod transaction;
 pub(crate) mod user;
