+++ backend/backend/src/api/routes/openapi/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod controller;
