+++ backend/tests/transaction/recurring.hurl
@@ -1,3 +1,10 @@
+# Get the id from the second currency
+GET {{api_v1}}/currency
+Authorization: Bearer {{admin_token}}
+HTTP 200
+[Captures]
+currency_id: jsonpath "$['data'][1]['id']"
+
 # Try to get all recurring transactions with out a token
 GET {{api_v1}}/transaction/recurring
 HTTP 401
@@ -20,7 +27,7 @@ Content-Type: application/json
   "description": null,
   "iban": null,
   "original_balance": 10000,
-  "currency_id": 2
+  "currency_id": {{currency_id}}
 }
 HTTP 201
 [Captures]
@@ -34,7 +41,7 @@ Content-Type: application/json
   "source_id": null,
   "destination_id": {{test_acc_id}},
   "amount": 1000,
-  "currency_id": 2,
+  "currency_id": {{currency_id}},
   "name": "10€ Deposit",
   "description": "A longer description bla bla",
   "budget_id": null
@@ -111,7 +118,7 @@ Content-Type: application/json
   "source_id": null,
   "destination_id": {{test_acc_id}},
   "amount": 1000,
-  "currency_id": 2,
+  "currency_id": {{currency_id}},
   "name": "10€ Deposit",
   "description": "A longer description bla bla",
   "budget_id": null
