+++ backend/tests/transaction/transaction.hurl
@@ -1,3 +1,10 @@
+# Get the id from the second currency
+GET {{api_v1}}/currency
+Authorization: Bearer {{admin_token}}
+HTTP 200
+[Captures]
+currency_id: jsonpath "$['data'][1]['id']"
+
 # Try to get all transaction without supplying a token
 GET {{api_v1}}/transaction
 HTTP 401
@@ -20,7 +27,7 @@ Content-Type: application/json
   "description": null,
   "iban": null,
   "original_balance": 10000,
-  "currency_id": 2
+  "currency_id": {{currency_id}}
 }
 HTTP 201
 [Captures]
@@ -35,7 +42,7 @@ Content-Type: application/json
   "description": null,
   "iban": null,
   "original_balance": 10000,
-  "currency_id": 2
+  "currency_id": {{currency_id}}
 }
 HTTP 201
 [Captures]
@@ -97,7 +104,7 @@ Content-Type: application/json
   "source_id": {{test_acc1_id}},
   "destination_id": {{test_acc2_id}},
   "amount": 1000,
-  "currency_id": 2,
+  "currency_id": {{currency_id}},
   "name": "10€ Transfer",
   "description": null,
   "budget_id": null,
