+++ backend/utility/benches/snowflake.rs
@@ -1,6 +1,5 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
-
-use utility::snowflake::SnowflakeGenerator;
+use utility::snowflake::generator::SnowflakeGenerator;
 
 pub fn next_id_benchmark(c: &mut Criterion) {
     let generator = SnowflakeGenerator::new(1, 0).expect("Failed to create SnowflakeGenerator");
